update: 1회성 타입 Interface를 파일 안으로 이동
