+++ src/components/MainPage/BoardItem.tsx
@@ -1,6 +1,10 @@
-import { type ListProps } from "@/types/board";
 import { cn } from "@/utils/cn";
 import { formatDateForList } from "@/utils/formatDate";
+
+export interface ListProps {
+	content: string;
+	createdAt: string;
+}
 
 const BoardItem = ({ content, createdAt }: ListProps) => {
 	const date = formatDateForList(createdAt);
