--- src/types/board.ts
@@ -1,4 +1,0 @@
-export interface ListProps {
-	content: string;
-	createdAt: string;
-}
