--- src/types/formatDateTypes.ts
@@ -1,8 +1,0 @@
-export interface ParseDateType {
-	year: number;
-	shortYear: string;
-	month: string;
-	day: string;
-	hour: string;
-	min: string;
-}
